move downloader impls to downloader module
